Add auth / unexpected error to AggregateException
Motivation
----------
In CarrierPublicationProvider, some exceptions are just logged despite
the fact that an AggregateException can be thrown without them being
listed in the aggregation of exceptions. Authentifcation and unexpected
errors are just logged, but if an AggregateException is thrown after,
it will not contain the auth/unexpected error.

Modifications
-------------
Both log and add the authentification / misc exceptions to the list of
aggregated exceptions, so that if an AggregateException is actually
thrown, it will allow to trace these errors.

Result
------
If an AggregateException is thrown after an authentication error or an
unexpected Exception, these are added to the list of cause exceptions.
If no AggregateException is thrown, this modification has no effect.

Change-Id: Id741fd561bda35ab9bf9286676138431d3e86120
Reviewed-on: http://review.couchbase.org/44575
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>