Allow JSON to be retrieved as a string
Motivation
----------
In certain scenarios, when a POCO is stored as JSON, you may want to
retrieve the string representation of that JSON. This patch allows you to
retrieve the value of a key that was stored as a JSON object as a string.

Modifications
-------------
The DefaultTranscoder was modified so that if the type of T is a string
and the common flag value is JSON, the client will deserialize the value
as a string.

Result
------
You can now retrieve a value stored as JSON and convert it to a string.

Change-Id: Iecc0b12ff9dcdfcff7ca330f0b0dcf8afe9988d2
Reviewed-on: http://review.couchbase.org/44604
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>