NCBC-753: Fix TimeSpan.ToTtl and test
Motivation
----------
Test in TimeSpanExtensionsTests was failing.

Modifications
-------------
This allowed to fix the method: the "<=" needed to be used to choose
the correct conversion (timestamp strictly over 30 days).

Additionaly in the test "Greater" assertion was used instead of "Less".

Result
------
The test passes and the conversion to timestamp is correctly only done
when strictly over 30 days.

Change-Id: I5d018957522d97ff994c2f2bdb119303529faa66
Reviewed-on: http://review.couchbase.org/44610
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>