Fix for improper buffer length in Operationbase.GetMessage
Motivation
----------
In most cases, the buffer returned after executing an operation will be a
24 byte header + the length of the body. When reading the body, you take
the length-header; in certain situations, you want to read the whole
buffer as the message.

Modifications
-------------
The OperationBase.GetMessage method was changed so that it would read the
whole buffer into the message field if the header length is shorter than
the buffer length.

Result
------
The message returned by the operation will be included in the Message
field, instead of the exceptions message (that would otherwise be thrown).

Change-Id: I56281a46a7c6c9d444a4c5317d59b9acf08cf70d
Reviewed-on: http://review.couchbase.org/45006
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>