Make "Percentage" field a string to avoid errors when value is "undefined"
Motivation
----------
The cluster may return a value of "undefined" for the Percentage fields.
The corresponding CLR type for these fields is an int, so it causes a
runtime exception.

Modifications
-------------
DatabaseFragmentationThreshold.Percentage and
ViewFragmentationThreshold.Percentage properties have been changed from
int to string.

Result
------
When the server sends a cluster map with values for Percentage that is
"undefined", no exception is thrown.

 verbose 'undefined' value valid on server

Change-Id: I8e91260023b6baa0f51ec0cc8539a793f6b02689
Reviewed-on: http://review.couchbase.org/45007
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>