NCBC-766: change client so that does not depend upon response body for Auth
Motivation
----------
Prior to this commit, the client would check the body of the SASL Step
response to determine if authentication was successful. This patch changes
the behavior to only check the response header for success.

Modifications
-------------
The Couchbase.Authentication.CramMD5AuthenticationMechanism class was
changed so that it only checks the response header for verification of
authentication.

Result
------
The client will only check the response header of a SaslStep (0x22)
request for authentication success or failure. The body will be ignored.

Change-Id: Ia320df587dbf25b85e4a6c5f1d96cdcfd677cc25
Reviewed-on: http://review.couchbase.org/44926
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>