NCBC-775: Unhandled Socket Exception during swap, failover, or rebalance
Motivation
----------
During swap/failover and rebalance scenarios, the cluster map will
often change resulting in the connections being closed locally or the
server itself will terminate the connection. In these cases it's probable
that operations may be in mid-flight. If the operation happens to be
occuring on a connection that was terminated, a runtime exception will be
thrown. This patch fixes a bug where a run-time exception is thrown and
not correctly handled, terminating the process the client is running in.
In general, operations should never result in uncaught exception; instead
the exception should be caught and associated with IOperationResult that
repsents the result of the operation so that application code can handle
the error as they please.

Modifications
-------------
Try/catch blocks were added to the DefaultIoStrategy and more importantly
in the OnCompleted handler for each Connection.

Result
------
When an exception is thrown during an asynchronous IO event, it will be
handled, logged, and finally associated with the IOperationResult of the
pending operation.

Change-Id: Iac449f3b72b00fa39d75d7365ff7b66df78ec982
Reviewed-on: http://review.couchbase.org/44945
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>