NCBC-777: Observe returns ObserveResponse.DurabilitySatisfied when CAS has changed
Motivation
----------
When observing a key, if the CAS has changed, the observe response should
be DurabilityNotSatisfied indicating a failed observe state.

Modifications
-------------
The CAS comparison is now done before the Durability requirements are
checked.

Result
------
If the CAS value of a key is changed, observe will return
ObserveResponse.DurabilityNotSatisfied, otherwise it will return
ObserveResponse.DurabilitySatisfied.

Change-Id: I18ca61fef080bf2eb25938432f8233daaf85de51
Reviewed-on: http://review.couchbase.org/44989
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>