NCBC-762: Make AutoCompactionSettings and FastWarmupSettings objects
Motivation
----------
On the IBUcketConfig interface and it's implementation, the
AutoCompactionSettings and FastWarmSettings are boolean, which causes a
parsing exception if they exist. This makes the fields the correct object
type so that serialization succeeds.

Modifications
-------------
IBucket.AutoCompactionSettings was changed from a bool to the
AutoCompactionSettings type, the same for IBucket.FastWarmupSettings.

Result
------
When these objects are contained within the config JSON, serialization
will succeed.

Change-Id: I3af60708b5ddf51784fb991f57b643f6223c146c
Reviewed-on: http://review.couchbase.org/45043
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>