Ensure Connection does not throw exception while disposing
Motivation
----------
When the cluster map changes and the connection pool needs to be
recreated, each connection is disposed. This patch ensures that an
exception thrown in the Dispose method is caught and logged at
LogLevel.INFO.

Modifications
-------------
Logging was added to BufferAllocator. A try/catch clause was added to the
Connection.Dispose method. The MRE.Set line was moved below the point were
the connection sets the State to exception.

Result
------
Exceptions thrown by Dispose are caught and logged.

Change-Id: I996197cfc568921480028821be88adecbb609fa5
Reviewed-on: http://review.couchbase.org/45092
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>