NCBC-770: Remove AutoCompactionSettings and FastWarmupSettings from IBucketConfig
Motivation
----------
In some cases the server returns a boolean value for these config elements
and at other times it sends an object. During serialization, the client
will throw an exception of it receives a type that it doesn't expect.
Since the client does not use these settings, they can be removed.

Modifications
-------------
AutoCompactionSettings and FastWarmupSettings from IBucketConfig
have been removed from IBucketConfig and it's implementations.
Additinally, they have been removed from Pools.

Result
------
The client will not throw Newtonsoft.Json.JsonSerializationException's for
these elements when the Type is not the expected Type. Note that the
client does not use these values.

Change-Id: I4046d477c9f98ba4e7f5c709b9ceb1aae594413c
Reviewed-on: http://review.couchbase.org/45115
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>