NCBC-734: Ensure password is used during post failover authentication
Motivation
----------
After and during a failover, the client will recieve configs from the
cluster via CCCP. When this happens, it needs to pick up the crendentials
from the client configuration, so that it can perform auth. This
patch fixes a bug were that crendentials were not being populated
resulting in failed authentication.

Modifications
-------------
ClusterCarrierProvider.GetConfig now stores the password in the client
bucket configuration and sets the password or the incoming server
configuration.

Result
------
Post failover, the client can authenticate with the server.

Change-Id: I483950d1a4b03f635e265d8942291e82fdec48c3
Reviewed-on: http://review.couchbase.org/45151
Reviewed-by: Andrei Baranouski <andrei.baranouski@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>