NCBC-784: get ClusterInfo via ClusterManager
Refactor ClusterInfo retrieval in order to go through the ClusterManager
rather than expose an API directly on Cluster. The rationale for this
change is that credentials are necessary to get the ClusterInfo, and
previous "alternative" of getting it through a cached configuration in
the provider would only work with Http configuration provider, making it
a clutch alternative.

Change-Id: I718279fb2aa35f968293e0635cf3df03e3423bf0
Reviewed-on: http://review.couchbase.org/45231
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>