NCBC-778: Ensure byte[] arrays stored in 1.3.X are correctly decoded in 2.0
Motivation
----------
When a value is stored as a byte[] array in 1.3.X and the key is retrieved
with 2.X, the value should be decoded correctly as a byte[] and not as a
JSON value.

Modifications
-------------
DefaultTranscoder was modified so that if the type is a byte[], it will be
decoded correctly as a byte.

Result
------
Byte[] arrays stored in 1.3.X can be read by 2.X clients correctly as the
same type: byte[].

Change-Id: I169661f66b2d6deb96bd17332a8950c985d1f5a8
Reviewed-on: http://review.couchbase.org/45285
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>