NCBC-749: Update Common.Logging to version 3.0 & Log4Net to 1.2.13.0
Motivation
----------
This is motivated by changes made by the Common.Logging project team with
respect to packaging:
https://github.com/net-commons/common-logging/wiki/Common.Logging-Packaging-and-Versions.
In order to support the new packaging scheme, the version of
Common.Logging had to be updated as well as the dependencies on Log4Net.
Note that the version of Common.Logging.Log4Net the SDK was using was
deprecated on NuGet, thus causing confusion with the user community.

Modifications
-------------
The Common.Logging version was updated from 2.0.0 to 3.0.0. The dependency
on Log4Net was updated from 1.2.10 to 1.2.13.0 and the
Common.Logging.Log4Net dependency was updated to
Common.Logging.Log4Net1213.

Result
------
The Common.Logging and it's dependent packages have been updated to the
lastest versions; users will no longer have to reference the deprecated
versions of these packages.

Change-Id: I15cbd8df7f2a077758d0a8a8836dc233835f7fb4
Reviewed-on: http://review.couchbase.org/45348
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>