NCBC-793: Make config thread for HttpStreamingProvider a background thread
Motivation
----------
If cluster object is properly disposed on a MemcachedBucket or
CouchbaseBucket where CCCP has failed, the configuration thread for
HttpStreamingProvider will hang indefinitly, since it's a foreground
thread.

Modifications
-------------
The configuration thread in ListenForConfigChanges is now a background
thread. A new MemcachedBucketTests_Isolated class was added for tests in
which the usual SetUp/TearDown methods should not be run.

Result
------
The process will no longer hang if the configuration thread has not
properly been shutdown by the Dispose method.

Use a background thread for ListenForConfigChanges so that the main
process isn't prevented from terminating.

Change-Id: I72b9e6338da313b46d24cc4124f027b780cb12c6
Reviewed-on: http://review.couchbase.org/45825
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>