NCBC-785: Get and GetDocument performance is slow when reading large documents
Motivation
----------
A bug was causing the read buffer to be the length of the request; if a
large document was read, it would take many reads to buffer it. This
commit fixes that issue and makes the BufferSize configurable so that if
the default 16K buffer is too small/large it can be tuned.

Modifications
-------------
The Connection class now uses the length of the buffer if the request body
is longer than the BufferSize. A property was added, BufferSize, which
allows the default 16k buffer to be increased to a desired size.

Result
------
While reading large documents, the client will no longer use the length of
the request buffer as the response buffer.

Change-Id: Ifa4a4b32261b83ac8a96953555564736cbc15e43
Reviewed-on: http://review.couchbase.org/45416
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>