NCBC-786: Add GetBucket(nm), GetBucket(nm, pw), RemoveBucket(nm) to ClusterHelper
Motivation
----------
Managing cluster and bucket instances can be tricky. These set of
overloads, allow a caller to open a bucket and cache it internally for
future use. When a subsequent request is made for a bucket by the same
name is made, the cached instance is returned.

Modifications
-------------
Three new methods were added to ClusterHelper: GetBucket(name),
GetBucket(name, pw) and RemoveBucket(name). These methods maintain references to the IBucket
instance that they return.

Result
------
Users of the SDK will have an easier interface to work with when using the
SDK in environments such as ASP.NET, which require bucket caching for
performance reasons.

Change-Id: If4952ff1f0100dbb822bc71716620e929d54f419
Reviewed-on: http://review.couchbase.org/45579
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>