NCBC-789: Add IsOpen(bucketName) to ICluster
Motivation
----------
In certain cases, such as testing and reference managment, it is useful to
know whether or not the Cluster object has a reference to an open IBucket
implementation. This commit adds a method called IsOpen(bucketName) which
returns true if the current Cluster is maintaining a reference to a
particular bucket.

Modifications
-------------
A method, bool IsOpen(string bucketName) was added to  the ICluster
interface. The internal IClusterController had a method added as well
which returns a value indicating if the bucket reference is being
observed.

Result
------
You can now determine whether or not a bucket is currently open from the
Cluster object.

Change-Id: Idd8f34e3831643ab6aae660a3d4cc9114ccdf0ff
Reviewed-on: http://review.couchbase.org/45580
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>