NCBC-794: Send documents larger than configured BufferSize
Motivation
----------
The buffer size for sending and receiving documents can be configured
using the ClientConfiguration.BufferSize field. However, if the document
exceeds this size, the client should still be able to send the document.
This patch enables the client to send documents larger than it's
configured buffer size.

Modifications
-------------
The SocketAsyncState class now has a field for storing the number of bytes
sent. The Connection class now uses this field to determine if it should
call Send(e) more than once.

Results
-------
Documents that are larger than the configured BufferSize can be sent to
the server.

Change-Id: I865260cd9e491139e2044002171e12e0f5680b2b
Reviewed-on: http://review.couchbase.org/45759
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>