NCBC-791: Ensure multi methods do not throw exceptions when keys are empty
Motivation
----------
If an empty set of keys are passed to a multi-method(Get, Upsert, Remove),
the method should simply return back an empty set, not throw and
ArgumentException.

Modifications
-------------
Before creating the partitionar, the multi-methods in CouchbaseBucket and
MemcachedBucket now check the size of the parameters and return an empty
set when the count is zero.

Change-Id: I3916df7f2153abcd619cf8f32a31cd2b4e66b88d
Reviewed-on: http://review.couchbase.org/45764
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>