Make noop use GetLengthSafe on Key
Motivation
----------
The noop operation overrides a base method passing in null values
resulting in an NullReferenceException for the Key field.

Modifications
-------------
Change noop to use GetLengthSafe instead of GetLength so that NRE is no
thrown if Key is null (which it always will be for a noop).

Result
------
A NullReferenceException will not be thrown.

Change-Id: Ibe510bb2b21cf2bf1013521a88d2e6641a9350a3
Reviewed-on: http://review.couchbase.org/45804
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>