NCBC-797: Client throws JsonSerializationException on Mono during bootstrapping
Motivation
----------
Using Mono and a non-Windows platform, the client will throw an
NotImplementedException which will be rethrown by NewtonSoft as a
JsonSerializationException. This is a Mono bug in that the getter is not
implemented:
http://mono.1490590.n4.nabble.com/Not-implemented-exceptions-in-Getters-td4663304.html

Modifications
-------------
In Cluster.LogConfigurationAndVersion the exception is caught and logged
under INFO.

Result
------
The exception will no longer be thrown by the client.

Change-Id: I9fbb4a0342c8deb8ab5d780dc0bce523fc9bdf48
Reviewed-on: http://review.couchbase.org/45951
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Patrick Varley <patrick@couchbase.com>