NCBC-800: fix encoding of query params and test
Motivation
----------
Query spec sometimes specifies that a parameter must be JSON, but most
of them are not. Despite that, some (like compression) are encoded in
QueryRequest.cs as JSON then urlencoded.

Scan_consistency is also badly rendered in the GET url, for example the
REQUEST_PLUS consistency is rendered as RequestPlus, which is rejected
by the server.

Modifications
-------------
Parameters that should be plainly encoded are not serialized to JSON
anymore, parameters that should actually be JSON encoded are (except
for scan_vector which haven't been tested).

Tests have been put in place to check that this encoding is well done
in GetFormValues(), that GetQueryParameters() is a simple appending of
GetFormValues() and that GetRequestUri()'s http parameters corresponds
to GetQueryParameters() output.

Result
------
Query parameter encoding better following N1QL DP4 specifications.

Change-Id: I1d94bcbc2364c5210e61ccd9a6d3f7db9a5c1c03
Reviewed-on: http://review.couchbase.org/46141
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>