NCBC-619: Make Socket connection timeout configurable
Motivation
----------
The default socket connect timeout is around 20s, so if the client tries
to connect to a node that's been removed from the cluster, then the app
will wait this time before failing. This commit adds a new property called
ConnectTimeout, which allows you to change this to a lower timespan. If
you set it higher than the Socket's default, it will honor the socket's
default and ignore the ConnectTimeout value.

Modifications
-------------
The ClientConfiguration has a new property, ConnectTimeout. The
DefaultConnectionFactory was nmodified to force a timeout based upon the
value of the ConnectTimeout. The default was set to 10s (this may change
in later releases).

Results
-------
If set lower than the Socket's default timeout of 20s, the ConnectTimeout
will force the client to abort a connection if it cannot be established by
the value defined by ConnectTimeout.

Change-Id: I00999cf644e5541e034927d1dd06091301322e0b
Reviewed-on: http://review.couchbase.org/46185
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>