NCBC-774: Ensure .NET SDK CAS behaviour is consistent with Java SDK 2.0
Motivation
----------
We want to ensure that the behavior of the IBucket methods which operate
on an IDocument<T> use CAS consistently between Java and C#.

Modifications
-------------
IBucket.Replace(IDocument<T>) now uses CAS, Insert and Upsert do not. Any
other method which uses a Key/Value overload has not been changed.

Result
------
The Java and .NET SDK's have the same Replace behavior for CAS: they use
it; Insert and Upsert do not.

Change-Id: Ib197704b2559a41c93e405ab6af19a5e3d029a3a
Reviewed-on: http://review.couchbase.org/46296
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>