NCBC-807: ops/sec drop to 0 during rebalance on 2.5 CB
Motivation
----------
During a rebalance the client may receive many configurations, so many
that it will all of it's time reconfigging. This fix ensures when a config
change is discovered, the client will only update the VBucketServerMap if
that is the only change, otherwise it will do a full reconfiguration.

Modifications
-------------
The CouchbaseConfigContext was changed so that recreating the servers and
the vbucket maps is done seperately and only when needed. Additionally,
equals methods were added to make it easier to detect smaller config
changes.

Result
------
The client will be much more efficient with respect to creating and
destroying connections.

Change-Id: Id80840d5182cfecce906e253583ebd9ae8ea010c
Reviewed-on: http://review.couchbase.org/46481
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>