NCBC-810: Ensure NodeAdapter list uses order of ServersList for VBucket mapping
Motivation
----------
In CB 2.5, the nodes list order may change and be different from the
ServerList usually during a rebalance in. When this happens, the client
will recieve NMV's. This fix ensures that the nodes and serverlist are
ordered the same. In addition, it makes the client more sensitive to
config changes.

Modifications
-------------
Sorting was added to the creation of the NodeAdapters based upon the
ServerList. Doing this broke many unit tests, so it cascaded into a lot of
changes to small util and extension classes. Unit tests were added to
ensure VBucketForwardMaps were correctly being treated as a configuration
change.

Results
-------
There will be a reduction in NMV's and the client will rebound after a
rebalance. There is still work to be done here, however.

Change-Id: I8aaed8611d91338802b677ae833995cf7f6ae507
Reviewed-on: http://review.couchbase.org/46554
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>