Flag connection as dead so it's not reused after IO exception
Motivation
----------
If a connection throws an exception, we want to flag it as dead so that
it's not reused. The connection will be cleaned up in the connection pool
and not added back into the queue.

Modifications
-------------
Before rethrowing an exception the connection is marked as dead.

Result
------
The connection will not be reused; instead it will free up a spot in the
pool for a new connection to be created.

Change-Id: I633786f9afeabee89aee2455c8ca2e713d681d2a
Reviewed-on: http://review.couchbase.org/46564
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>