NCBC-805: use graceful pool teardown for dispose only
Motivation
----------
When the SocketPool is finalized, we do not want to block the finalization
thread, so we simply dipose of everything and suppress any unhandled
exceptions.

Modifications
-------------
The SocketPool class was modified so that if the Dispose method is called
during finalization, it will simply go through and attempt to dispose of
each connection. If an exception is thrown, it will be suppressed.

Result
------
If disposed is called by a non-finalization thread, the client will
attempt to close each socket using a backoff algorithm. However, if the
finalization thread calls Dispose, each socket will be closed without
regard to the status of the socket.

Change-Id: Ic684ee609f6e715560be228432eb30e75459135a
Reviewed-on: http://review.couchbase.org/46660
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>