NCBC-802: Rename ConnectionTimeout to SendTimeout
Motivation
-----------
For clarity and to align with other namings, rename ConnectionTimeout
to SendTimeout in PoolConfiguration. The tuning should also be made
possible at the App.config level.

Modifications
-------------
Renamed the parameter in PoolConfiguration.
Added a property for it in ConnectionPoolElement and let it be picked
up by bucket configurations when constructing the ClientConfiguration.

Added tests related to this parameter and ammended a few existing
ones.

Change-Id: I0211d44a5d01e6c462a463e33c6ec7a9dff2f082
Reviewed-on: http://review.couchbase.org/47076
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>