NCBC-824: Cluster.OpenBucket honor pwd in config
Motivation
----------
Same as NCBC-818 for ClusterHelper, Cluster.OpenBucket(name) should try
to use the password in ClientConfiguration.BucketConfigs if one is
found there.

Modifications
-------------
Cluster.OpenBucket(name) now peek into the ClientConfiguration and uses
the password for bucket name if it can be found. Otherwise default to
empty password (use OpenBucket(name, password) if you want to be sure).

Added related test.

Result
------
If there's a password defined for the bucket in configuration, it is
automatically picked up if no password is explicitely provided.

Change-Id: I3c98112c2ed660902e1e54283dba7443b1ca1d03
Reviewed-on: http://review.couchbase.org/47116
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>