NCBC-816: Rework prepared statement API
Motivation
----------
In order to have coherence between the Java SDK and the .Net SDK, a
rework of the Prepared Statement N1QL feature is needed. The new API
will not involve transparent caching of prepared queries but will leave
it to the user. The two steps of the process (preparing the plan and
executing it) will be made explicit in the API (cf Java SDK).

Modifications
-------------
Removed IPreparable and all references to it, removed all occurrences
of caching.
Added IQueryPlan interface. Added a method on IQueryClient to prepare
a statement and obtain the associated IQueryPlan.
Reworked IQueryRequest / QueryRequest to consider such a IQueryPlan the
only source of prepared request (IsPrepared = true).
Reworked actual query building and existing unit tests to reflect that.

Result
------
Doing a prepared query is now fully into the hands of the user who can
decide when to best cache the plan he can now directly obtain.
API is coherent with the Java SDK one.

Change-Id: I90ed2b47806ff110ec787e6c293d16067ebd8450
Reviewed-on: http://review.couchbase.org/47293
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>