Move hardcoded IPs in unit tests to config
Motivation
----------
Several of the unit tests use IPs and Uris that are hardcoded within the
tests which makes running the in environments other than localhost
difficult. This patch moves those values to the App.Config so that it is
easier to change them to run in an enviroment other than localhost. This
is one step of many required to automate UTs.

Modifications
-------------
In all tests if a hardcoded IP exists, they were moved to the App.Config.

Results
-------
To run the tests outside of localhost, change the OperationTestAddress
(ip+port) in the App.Config.

Change-Id: I335102581f0765e73694ddbdc19b5f04c20728d6
Reviewed-on: http://review.couchbase.org/47593
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>