Update management classes to use pools API
Motivation
-----------
The pools API gives all the information to construct the Cluster API
calls. This commit uses that information to construct the REST calls and
refactors/cleans up some duplicate code.

Modifications
-------------
The ClusterManager now takes takes an IServerCOnfig in it's ctor and uses
the Controllers values to construct the URI. The construction of the URI
is now done by one of two methods: GetApiURI and GetBucketApiUri. The
Controllers class was removed and replaced with a Controller class and the
Pools class modified to use a Dictionary of Controller classes from the
pools configuration.

Result
------
The Management API uses the correct source to generate it's URI's.
Duplicate code has been refactored and is in general, easier to
maintain.

Change-Id: Id2749a3576a90b80c9b8934274503bf407b61fd4
Reviewed-on: http://review.couchbase.org/47653
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>