NCBC-831: Make Prepare part of the Bucket API
Motivation
----------
QueryClient is just there for modularity and should be made internal.
Thus, the Prepare API that has previously been added should be instead
exposed in the IBucket interface.
Also, the use of QueryRequest parameters should be favored over uri and
string.

Modifications
-------------
Prepare API has been exposed in IBucket, delegating to IServer and at
last delegating to the QueryClient.
Both string and IQueryRequest overloads are available. In both cases
only the Statement is used, the IQueryRequest one will see its BaseUri
overwritten by the one of the Server.

IQueryClient has been fixed and now only take a IQueryRequest as input.
Out of this request, only the BaseUri and Statement are used in order
to obtain a IQueryPlan.

Tests have been added/fixed in consequence.

Result
------
A statement (either as plain string or embedded in a IQueryRequest) can
now be prepared from the IBucket.

Change-Id: Ia46f2df09fb265307f935cae41dda2b3af21c264
Reviewed-on: http://review.couchbase.org/48118
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>