NCBC-817: catch ServerUnavailable in IsSecure
Motivation
----------
In some cases (like IoC with Autofac for instance), the server cannot
be reached and the IsSecure fails with a ServerUnavailableException.
This should be guarded against, logging the error and returning false.

Modifications
-------------
CouchbaseBucket and MemcachedBucket IsSecure now catches the exception.
It is logged (info level) and the method defaults to returning false.

Result
------
Less irrelevant exception when using dependency injection.

Change-Id: I791ec7fae3b8dec82a64c322937b70a1e98a61d7
Reviewed-on: http://review.couchbase.org/48178
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>