NCBC-832: Refactor 'SendWithRetry' methods into IRequestExecuter
Motivation
----------
The IBucket implementations have retry logic strewn about - this ticket is
to refactory that logic into a more cohesive IRequestExecuter
implementation for Memcached and Couchbase buckets.

Modifications
-------------
A new interface, IRequestExecuter, has been defined with two
implementations: MemcachedRequestExecuter and CouchbaseRequestExecuter.
The CouchbaseBucket and MemcachedBuckets have been updated to use the
executer, however some work still needs to be done in a later commit for
async/await; this is depenendent on another changset, though.

Results
-------
The send and retry logic is now centrally located and easier to mock and
test.

Change-Id: If7f85329bf5cb974aa108ab8b8dead5b2970d24e
Reviewed-on: http://review.couchbase.org/48276
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>