NCBC-828: Provide support for GetAndTouch and Touch
Motivation
----------
GetAndTouch (GAT) and Touch allow for a keys expiration to be updated.
This is also known as "sliding expirations". While touch simply updates
the expiration of an exiting key, failing if it doesn't exist, GAT will
return the value of the key as well.

Modifications
-------------
The Touch and GetT commands were implemented. The CouchbaseBucket and
MemcachedBucket have had additional methods to support these commands
added. Support for methods which return IOperationResult and
IDocumentResult were added.

Results
-------
The SDK now supports these methods which existed in v1.3.X and were absent
for v2.X.

Change-Id: I1e96c043c4988e643b0858be13bda3def089a977
Reviewed-on: http://review.couchbase.org/48656
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>