NCBC-790: Add Support for Exists method
Motivation
----------
The Exists method was supported in 1.3.X versions of the SDK, but absent
from v2.X. This commit adds the functionality to the SDK.

Modifications
-------------
Two methods were added Exists and ExistsAsync were added to the IBucket
interfaces and implementations were created for Memcached and Couchbase
buckets. Note that the async versions are not implemented in this commit,
however the signatures were added. The impolementation for these methods
will be added in a subsequent commit after the async-await branch is
merged with master.

Results
-------
You can now check for the existence of a key using the Exist method. Async
support for ExistsAsync will be added in a later commit.

Change-Id: I264735c0b1fd85732f9b28a070b9757cfd567fd8
Reviewed-on: http://review.couchbase.org/48663
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>