NCBC-808: Append/Prepend/Incr/Decr use retry path
Motivation
----------
All operations, including Append, Prepend, Increment and Decrement, can
and should retry when a NotMyVBucket error is returned.

The introduction of the RequestExecuter, and the Couchbase bucket
implementation for it is consistent with this statement, however the
executer is not invoked by the CouchbaseBucket for these operations.

Modifications
-------------
Make CouchbaseBucket use the RequestExecuter's SendWithRetry method.
Still use GetServer to retrieve the vbucket so that it can be given to
the operation (since this is not a memcached bucket).

Removed dead code (CheckConfigForUpdate method in CouchbaseBucket).

Result
------
Increment/Decrement/Append/Prepend use the common path for operation
execution with retry support. They'll now be retryed in case of NMV in
Couchbase buckets.

Increment and Decrement are otherwise listed as non-retriable, whereas
Append and Prepend can be retried if the status permits it and Cas > 0

Change-Id: I7919eb6337e8fe8eadc584cd185fedf67b5de5a4
Reviewed-on: http://review.couchbase.org/48805
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>