NCBC-842: Add overloads for non-generic SendWithRetry and SendWithRetryAsync
Motivation
----------
Certain operations, Touch for example, do not have a body, thus the
generic Type parameters are meaningless. In the past, a placeholder Type
was used. In this commit we are changing the requirement for place holder
types by providing non-generic operations and execution flows.

Modifications
-------------
OperationBase<T> was refactored into two classes: a non-generic and a
generic version. Operations with no body inherit from the non generic one.
The Clone method was changed to be non-generic. The IRequestExecutor
interface and implementations, the IServer and it's implmentation, and
IIOstrategy were updated to support the non-generic operations as well.

Result
------
Placeholder types no longer need to be used for non-generic operations.
Specialization and cohesion have been improved.

Change-Id: I12de8cb3e4cc7e79767175dfdea08823ca19fbc0

Conflicts:
	Src/Couchbase/Core/Buckets/IRequestExecuter.cs
Change-Id: I33ac14d83a74c3540a3ef918e0658f18617f5604
Reviewed-on: http://review.couchbase.org/49142
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>