NCBC-845: NMV returned turned to caller during rebalance
Motivation
----------
During a swap/failover/rebalance an operation which encounters a NMV
should be retried until it times out. The SDK should never return a NMV to
the caller.

Modifications
-------------
A new method was added to the IOperationResult interface which determines
if an NMV was encountered. The IRequestExecutor implementations were also
refactored to ensure the correct logic was used.

Result
------
NMVs are no longer returned to the caller during rebalances and the SDK
will correctly retry until it either succeeds or times out.

Change-Id: I3e9e839cc7097da16090ffe6821ab8faea6103c9

Conflicts:
	Src/Couchbase/Core/Buckets/RequestExecuterBase.cs
Change-Id: I5afd9a5a32e2255106234e08d480cfbc18967aca
Reviewed-on: http://review.couchbase.org/49146
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>